venv-requirements.txt
Django==4.1.3
Pillow
pytz
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.5
PyJWT==2.1.0
# django-crispy-forms
crispy-bootstrap5
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
django-crum==0.7.9
django-filter==21.1
django-jazzmin==2.5.0
debugpy==1.6.2
requests==2.28.1
# supervisor
# uwsgi
# models UML
django-extensions
pyparsing
pydot
